PSR‑4 Hardening for Today Helpers
 PSR‑4 Hardening for Today Helpers: split TodayHelpers classes into separate files to align with Composer PSR‑4 and remove explicit requires
